Parallel Processing Update
The Parallel Downloader has been refactored and now takes advantage of
the Threaded::extend() method, which provides thread safety to the
CurlDownloadChunk at runtime. While this is a bad practice in most
cases, it doesn't seem to be optional in the context of pthreads, and
is the most performant way to perform parallel processing.